[1865] Refactor MCB Editors to share init and clearer folder structure
