+++ bin/mcb
@@ -22,11 +22,12 @@ require 'mcb'
 require 'mcb/azure'
 require 'mcb/config'
 require 'mcb/course_show'
-require 'mcb/courses_editor'
-require 'mcb/provider_editor'
 require 'mcb/cli/base_cli'
 require 'mcb/cli/course_cli'
 require 'mcb/cli/provider_cli'
+require 'mcb/editor/base'
+require 'mcb/editor/courses_editor'
+require 'mcb/editor/provider_editor'
 require 'mcb/grant_access_wizard'
 require 'mcb/revoke_access_wizard'
 require 'mcb/render'
