+++ lib/mcb/commands/providers/create.rb
@@ -7,7 +7,7 @@ run do |opts, _args, _cmd|
     provider = MCB.get_recruitment_cycle(opts).providers.build
     requester = User.find_by!(email: MCB.config[:email])
 
-    MCB::ProviderEditor.new(
+    MCB::Editor::ProviderEditor.new(
       provider: provider,
       requester: requester
     ).new_provider_wizard
